feat(sfml) : add sfml graphic lib (#31)
* feat(sfml): add SFML graphic library base structure

* feat(sfml): implement SFML display and renderer classes with event handling

* feat(scene): set default font size for text in scene

* feat(renderer): enhance sprite drawing with dynamic texture sizing

* feat(sfml): include additional headers for string and unordered_map support

* feat(menu): add functions to format library names for display selection

* feat(sfml): include string header for improved functionality in display and renderer

* feat(clangd): update include path to use Homebrew directory for compatibility

* feat(sfml): remove unnecessary include for IDisplay interface

* feat(sfml): add printable character constants and update event handling

* feat(sfml): add NOLINT comments for improved code clarity

* feat(sfml): enhance code clarity by using 'const' for variable declarations

* feat(sfml): remove src.list file as it is no longer needed

* feat: improve formating files

+++ src/lib/game/pacman/map.cpp
@@ -56,9 +56,10 @@ bool Pacman::loadMap(const std::string& filePath) {
 }
 
 type Pacman::getTile(int xCoordinate, int yCoordinate) {
-  if (xCoordinate < 0 || xCoordinate >= static_cast<int>(_map.size()) ||
+  if (xCoordinate < 0 || static_cast<std::size_t>(xCoordinate) >= _map.size() ||
       yCoordinate < 0 ||
-      yCoordinate >= static_cast<int>(_map[xCoordinate].size())) {
+      static_cast<std::size_t>(yCoordinate) >=
+          _map[static_cast<std::size_t>(xCoordinate)].size()) {
     return type::WALL;
   }
   return _map[xCoordinate][yCoordinate];
